requirements.txt
Django==6.0
cloudinary>=1.44,<2
django-ckeditor==6.7.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg[binary]>=3.2,<4
python-dotenv==1.2.1
requests==2.32.5
redis>=6,<7
uvicorn==0.40.0
whitenoise==6.11.0
